:recycle: Splitting up Resources.test into multiple files
