+++ .github/workflows/code_coverage.yaml
@@ -1,9 +1,15 @@
-name: Test
+name: Coverage
 on:
   workflow_call:
 jobs:
   code-coverage:
     runs-on: ubuntu-latest
+
+    permissions:
+      # Required to checkout the code
+      contents: read
+      # Required to put a comment into the pull-request
+      pull-requests: write
 
     steps:
       - uses: actions/checkout@v4
