+++ vitest.config.ts
@@ -4,7 +4,7 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   plugins: [viteTsconfigPaths() as any],
   test: {
-    pool: "threads",
+    maxConcurrency: 35,
     server: {
       deps: {
         inline: ['@equinor/subsurface-app-management']
